+++ src/evaluate_compression.rs
@@ -1,4 +1,11 @@
-use std::{io::Read, time::Instant};
+use std::{
+    io::Read,
+    sync::{
+        atomic::{AtomicUsize, Ordering},
+        Arc,
+    },
+    time::Instant,
+};
 
 use crate::{
     config::{
@@ -10,6 +17,7 @@ use crate::{
     send::util::file_with_bufreader,
 };
 use anyhow::{bail, Result};
+use rayon::iter::{IntoParallelIterator, ParallelIterator};
 use strum::IntoEnumIterator;
 
 pub mod compression_result;
@@ -23,6 +31,7 @@ pub fn evaluate_compression(args: EvaluateCompressionArgs) -> Result<()> {
         input_file,
         omit,
         mut omit_levels,
+        threads,
     } = args;
 
     omit_levels.sort_unstable();
@@ -97,12 +106,72 @@ pub fn evaluate_compression(args: EvaluateCompressionArgs) -> Result<()> {
         }
     }
 
-    let compression_results: Vec<CompressionResult<Finished>> = compression_awaiting
-        .into_iter()
-        .flat_map(|cr_await| cr_await.run(&test_contents).ok())
-        .collect();
+    log::info!(
+        "Evaluating {} compression combinations",
+        compression_awaiting.len()
+    );
+    if threads == 1 {
+        log::info!("Running sequentially on the main thread");
+    } else {
+        log::info!("Running sequentially with up to {threads} threads");
+    }
+    rayon::ThreadPoolBuilder::new()
+        .num_threads(threads)
+        .build_global()?;
 
-    print_results::evaluate_and_printout_results(&compression_results);
+    // Setup the progress bar
+    let res = single_progress_bar(compression_awaiting, &test_contents)?;
+
+    print_results::evaluate_and_printout_results(&res);
 
     Ok(())
+}
+
+fn single_progress_bar(
+    compression_awaiting: Vec<CompressionResult<Awaiting>>,
+    test_contents: &Vec<u8>,
+) -> anyhow::Result<Vec<CompressionResult<Finished>>> {
+    use indicatif::style::*;
+    use indicatif::*;
+    let total = compression_awaiting.len();
+    let pb = ProgressBar::new(total as u64);
+    pb.set_style(
+        ProgressStyle::with_template(
+            "{spinner:.green} [{elapsed_precise}] [{bar:40.cyan/blue}] {pos:>7}/{len:7} ({eta})",
+        )?
+        .progress_chars("##-"),
+    );
+
+    // Atomic counter for tracking progress in the parallel iterator
+    let progress_counter = Arc::new(AtomicUsize::new(0));
+
+    let res = std::thread::scope(|s| {
+        let res = s.spawn(|| {
+            let compression_results: Vec<CompressionResult<Finished>> = compression_awaiting
+                .into_par_iter()
+                .flat_map(|cr_await| {
+                    let compr_res = cr_await.run(test_contents).ok();
+                    if let Some(ref compr_res) = compr_res {
+                        pb.suspend(|| {
+                            log::info!(
+                                "{}\n{}",
+                                compr_res.compression_format(),
+                                compr_res.summarize_as_table()
+                            );
+                        })
+                    }
+                    let current_progress = progress_counter.fetch_add(1, Ordering::SeqCst);
+                    pb.set_position(current_progress as u64 + 1);
+                    compr_res
+                })
+                .collect();
+            compression_results
+        });
+
+        let res = res.join().unwrap();
+        pb.finish_with_message("Processing complete");
+        res
+    });
+
+    Ok(res)
 }